Fix internal server errors
Seems like Redis does not return a null response if a key does not exist
but instead it returns a default value (e.g. all fields null in HMGET).

Additionally, I assumed for some reason that every software has a
metrics class and an example plugin. This is not true.