Add bankruptcy protection
SaaS with a pay-as-you-go model like Firestore is very scary because a
single mistake (e.g., an infinite loop) can easily accumulate an
enormous amount of writes or reads that may cost you a shit ton of money
in a very short time.

This commits adds a manual ratelimit by incrementing a counter in Redis
on every firestore write or read. Not perfect, but sadly Google does not
provide a better solution